Macro sscanf::sscanf

sscanf!() { /* proc-macro */ }

A Macro to parse a string based on a format-string, similar to sscanf in C

Signature

sscanf!(input: impl Deref<Target=str>, format: <literal>, Type...) -> Result<(Type...), sscanf::Error>

Parameters

input: The string to parse. Can be anything that implements Deref<Target=str> (e.g. &str, String, Cow<str>, etc. See examples below). Note that sscanf does not take ownership of the input.

format: A literal string. No const or static allowed, just like with format!().

Type...: The types to parse. See Custom Types for more information.

Return Value

A Result with a tuple of the parsed types or a sscanf::Error. Note that an error usually indicates that the input didn’t match the format string, making the returned Result functionally equivalent to an Option, and most applications should treat it that way. An error is only useful when debugging custom implementations of FromStr or FromScanf. See sscanf::Error for more information.

Details

The format string has to be a string literal (with some form of " on either side), because it is parsed by the procedural macro at compile time and checks if all the types and placeholders are matched. This is not possible from inside a variable or even a const &str somewhere else.

Placeholders within the format string are marked with {}. Any '{' or '}' that should not be treated as placeholders need to be escaped by writing '{{' or '}}'. For every placeholder there has to be a type name inside the {} or exactly one type in the parameters after the format string. Types can be referenced by indices in the placeholder, similar to format!().

Any additional formatting options are placed behind a :. For a list of options, see the crate root documentation.

Examples

A few examples for possible inputs:

let input = "5"; // &str
assert_eq!(sscanf!(input, "{usize}").unwrap(), 5);

let input = String::from("5"); // String
assert_eq!(sscanf!(input, "{usize}").unwrap(), 5);

let input = &input; // &String
assert_eq!(sscanf!(input, "{usize}").unwrap(), 5);
assert_eq!(sscanf!(input.as_str(), "{usize}").unwrap(), 5);

let input = std::borrow::Cow::from("5"); // Cow<str>
assert_eq!(sscanf!(input, "{usize}").unwrap(), 5);

let input = std::rc::Rc::from("5"); // Rc<str>
assert_eq!(sscanf!(input, "{usize}").unwrap(), 5);

// and many more

// temporary value: doesn't work
sscanf!(String::from("5"), "{usize}");

More Examples can be seen in the crate root documentation.