Macro sscanf::sscanf_get_regex

sscanf_get_regex!() { /* proc-macro */ }

Same as sscanf, but returns the regex without running it. Useful for debugging or efficiency.

Signature

sscanf_get_regex!(format: <literal>, Type...) -> &'static Regex

Parameters

format: A literal string. No const or static allowed, just like with format!().

Type...: The types to parse. See Custom Types for more information.

Returns: A reference to the generated Regex.

The Placeholders can be obtained by capturing the Regex and using the 1-based index of the Group.

Examples

use sscanf::sscanf_get_regex;
let input = "Test 5 -2";
let regex = sscanf_get_regex!("Test {usize} {i32}");
assert_eq!(regex.as_str(), r"^Test (\+?\d{1,20}) ([-+]?\d{1,10})$");

let output = regex.captures(input);
assert!(output.is_some());
let output = output.unwrap();

let capture_5 = output.get(1);
assert!(capture_5.is_some());
assert_eq!(capture_5.unwrap().as_str(), "5");

let capture_negative_2 = output.get(2);
assert!(capture_negative_2.is_some());
assert_eq!(capture_negative_2.unwrap().as_str(), "-2");