Trait sscanf::RegexRepresentation

pub trait RegexRepresentation {
    const REGEX: &'static str;
}

A Trait used by sscanf to obtain the Regex of a Type

Has one associated Constant: REGEX, which should be set to a regular Expression. Implement this trait for a Type that you want to be parsed using sscanf.

The Regular Expression should match the string representation as exactly as possible. Any incorrect matches might be caught in the from_str parsing, but that might cause this regex to take characters that could have been matched by other placeholders, leading to unexpected parsing failures.

Implementing the Trait

A manual implementation of this trait is only necessary if you

want to use a custom Type that is not supported by default AND

cannot use #[derive(FromScanf)] on your Type

Deriving FromScanf will automatically implement this trait for your Type, and should be preferred in most cases.

If you do need to implement this trait yourself, note the following:

The regex cannot contain any capture groups (round brackets). If you need to use ( ) in your regex, use (?: ) instead to make it non-capturing.

Using a raw string literal (r"...") is recommended to avoid having to escape backslashes.

The [const_format] crate can be used to combine multiple strings into one, which is useful for complex regexes. This can also be used to combine the existing regex implementation of other types. sscanf internally uses const_format as well, so a version of it is re-exported under sscanf::const_format.

Example

Let’s say we want to add a Fraction parser

struct Fraction(isize, usize);

Which can be obtained from any string of the kind ±X/Y or just X

impl sscanf::RegexRepresentation for Fraction {
    /// matches an optional '-' or '+' followed by a number.
    /// possibly with a '/' and another Number
    const REGEX: &'static str = r"[-+]?\d+(?:/\d+)?";
    //                                     ^^ escapes the group. Has to be used on any ( ) in a regex.

    // alternatively, we could use const_format to reuse existing regexes:
    // REGEX = const_format::concatcp!(isize::REGEX, "(?:", "/", usize::REGEX, ")?");
}
impl std::str::FromStr for Fraction {
    type Err = std::num::ParseIntError;
    fn from_str(s: &str) -> Result<Self, Self::Err> {
        let mut iter = s.split('/');
        let num = iter.next().unwrap().parse()?;
        let mut denom = 1;
        if let Some(d) = iter.next() {
            denom = d.parse()?;
        };
        Ok(Fraction(num, denom))
    }
}

Now we can use this Fraction struct in sscanf:

let output = sscanf!("2/5", "{}", Fraction);
assert_eq!(output.unwrap(), Fraction(2, 5));

let output = sscanf!("-25/3", "{}", Fraction);
assert_eq!(output.unwrap(), Fraction(-25, 3));

let output = sscanf!("8", "{}", Fraction);
assert_eq!(output.unwrap(), Fraction(8, 1));

let output = sscanf!("6e/3", "{}", Fraction);
assert!(output.is_err());

let output = sscanf!("6/-3", "{}", Fraction);
assert!(output.is_err()); // only first number can be negative

let output = sscanf!("6/3", "{}", Fraction);
assert_eq!(output.unwrap(), Fraction(6, 3));

Required Associated Constants

const REGEX: &'static str

A regular Expression that exactly matches any String representation of the implementing Type